// src/app/models/product.model.ts
// Generated by https://quicktype.io

export interface Product {
  idProduct?:      number;
  productName:    string;
  key:            string;
  price:          number;
  productStatus:  boolean;
  typeProduct:    TypeProduct;
  suppliers?:      Supplier[];
}

export interface TypeProduct {
  idTypeProduct:   number;
  typeProductName: string;
  typeProduct:     string;
}

// Generated by https://quicktype.io

export interface Supplier {
  idSupplier:   number;
  supplierName: string;
}

export interface SupplierProductAdd {
  idProductSupplier?: number | null;
  idProduct?: number;
  productSupplierStatus?: boolean;
  idSupplier: number;
  supplierName: string;
  supplierProductKey: string;
  supplierCost: number
  /**
   * Complete Product Object
   */
  product?: Product;
  /**
   * Complete Supplier Object
   */
  supplier?: Supplier;
}